+++ types/index.d.ts
@@ -1,3 +1,4 @@
+import { VueConstructor } from 'vue/types/umd';
 import { CombinedVueInstance, Vue } from 'vue/types/vue';
 
 export interface UploaderProp {
@@ -84,19 +85,24 @@ export interface UploaderMethod<ImageType> {
      */
     transformImage(image: ImageType): string;
 }
-export interface WechatImage {
+export interface BasicImage {
+    key?: string;
+    url: string;
+}
+export interface WechatImage extends BasicImage {
+    /** @deprecated */
+    image: string | null;
+    localId?: string | null;
     serverId?: string | null;
-    image: string;
-    base64?: string;
 }
 /**
  * InputUploader中所存储的内容
  */
-export interface FileImage {
+export interface FileImage extends BasicImage {
     /** file文件 */
     file?: File;
     /** 由file转换成的objectURL对象 */
-    objectUrl: string;
+    objectUrl?: string;
 }
 export interface WechatUploaderProp {
     useWechatPreview: boolean;
@@ -111,14 +117,17 @@ export interface WechatUploaderMethod {
      */
     uploadWechatImages(localIds: string[]): Promise<void>;
 }
+export interface InputUploaderProp {
+    accept?: string;
+}
 export interface UploaderOptions {
     transformWXLocalImageData: boolean;
 }
 
 export type UploaderComponnet = CombinedVueInstance<Vue, UploaderData<string>, UploaderMethod<string>, object, UploaderProp>;
 export type WechatUploaderComponent<T extends WechatImage = WechatImage> = CombinedVueInstance<Vue, UploaderData<T>, UploaderMethod<T> & WechatUploaderMethod, object, UploaderProp & WechatUploaderProp>;
-export type InputUploaderComponent = CombinedVueInstance<Vue, UploaderData<FileImage>, UploaderMethod<FileImage>, object, UploaderProp>;
-export const Uploader: UploaderComponnet;
-export const WechatUploader: WechatUploaderComponent;
-export const InputUploader: InputUploaderComponent;
+export type InputUploaderComponent = CombinedVueInstance<Vue, UploaderData<FileImage>, UploaderMethod<FileImage>, object, UploaderProp & InputUploaderProp>;
+export const UploaderFactory: (Vue: any) => VueConstructor;
+export const WechatUploaderFactory: (Vue: any) => VueConstructor;
+export const InputUploaderFactory: (Vue: any) => VueConstructor;
 export declare function install(vue: typeof Vue, options: UploaderOptions): void;