+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.3.0
+
+1. feat: WechatImage和InputImage均补充BasicImage，基本向后兼容，但WechatImage中的image不应该再使用
+1. feat: 补充导出 UploaderFactory / InputUploaderFactory / WechatUploaderFactory
+1. fix: 删除错误的导出类型 Uploader / WechatUploader / InputUploader
+
 ### 0.2.3
 
 1. feat: 优化代码，删除async/await形式代码